DNW: Bug fix: don't pass copy of steps to CaseSteps table
The copy was being made in order to sort it (without mutating a prop).
And I thought that since the copy was being supplied to the downstream
component by a computed property that it would react to changes in the
source data, but that doesn't seem to be true. So, we'll have to ensure
that the data is already correctly sorted.