Gantt task order: break ties on duration
Tasks are sorted by onset time, with ties broken by duration. This has
the effect that if task B depends on task A, then task A will sort
earlier than task B (either because it starts earlier or, if it starts
on the same day, then because its duration is shorter). In practice, if
task B depends on task A, and they start on the same day, then task A
has a duration of zero days (it is some sort of "event" pseudo-task).